refactor(api): créer une factory de contrôleur pour updateComment refers to https://trello.com/c/evsOelCU/2717
Les contrôleurs actionComment.updateComment et shantytownComment.updateComment
étaient strictement identiques à l'exception du service appelé. Extraction
dans makeUpdateCommentController, une factory générique paramétrée par la
fonction de service, réutilisée par les deux contrôleurs. Comportement HTTP
inchangé (mêmes codes ERROR_RESPONSES, même statut 200, même forme de
réponse).
